+++ lib/models/money_objects/transactions/transactions.dart
@@ -20,7 +20,7 @@ class Transactions extends MoneyObjects<Transaction> {
 
     for (final MyJson row in rows) {
       final Transaction t = Transaction.fromJSon(row, runningBalance);
-      runningBalance += t.amount.value;
+      runningBalance += t.getNormalizedAmount();
       getList().add(t);
     }
     return getList();